requirements.txt
aiohttp == 3.8.4
aiohttp-jwt == 0.6.1
aiofiles == 23.1.0
alembic == 1.10.2
asyncpg == 0.27.0
bcrypt == 4.0.1
loguru == 0.6.0
pyjwt == 2.6.0
sqlalchemy == 2.0.6
pydantic == 1.10.4
python-dotenv == 0.21.1
pytz == 2022.7.1
